+++ db/migrate/20260101000022_change_all_comment_notices_default_to_true.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+class ChangeAllCommentNoticesDefaultToTrue < ActiveRecord::Migration[7.0]
+  def up
+    change_column_default :bible270_readers, :notify_on_all_comments, from: false, to: true
+  end
+
+  def down
+    change_column_default :bible270_readers, :notify_on_all_comments, from: true, to: false
+  end
+end